Refactor file manager. Break it down into files
