+++ src/sync/syncManager/index.ts
@@ -6,8 +6,7 @@ import { DiffManager } from '../diffManager';
 import { Renderer } from '~/renderer';
 import { diffBooks } from './diffBooks';
 import type FileManager from '~/fileManager';
-import type { KindleFile } from '~/fileManager';
-import type { Book, BookMetadata, Highlight } from '~/models';
+import type { Book, BookMetadata, Highlight, KindleFile } from '~/models';
 import type { DiffResult } from '../diffManager';
 
 export default class SyncManager {
